data-warehousing/pyproject.toml
[project]
name = "data-warehousing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-storage>=3.9.0",
    "google-cloud-bigquery>=3.0.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
]
